refactor(fixedIncome): merge Bond into CLBond with explicit constructor
Bond and CLBond are merged into a single CLBond class since the module
only ever targets the Chilean market (base-100 flows, TERA). CLBond's
constructor now takes explicit typed parameters with validation instead
of **kwargs, removing the double __init__ chain that existed only
because of the Bond/CLBond inheritance.

Also fixes a silent business-rule check: Coupons.adjust_to_notional used
an assert (disabled under python -O) for its residual-sum sanity check;
it now raises ValueError unconditionally.
