+++ src/fintoolsom/__init__.py
@@ -1,7 +1,7 @@
 from importlib.metadata import version as _v
 
 __version__ = _v("fintoolsom")
-from .fixedIncome import CLBond, Bond, Coupons, Coupon
+from .fixedIncome import CLBond, Coupons, Coupon
 from .fixedIncome import get_irr
 from .models import NelsonSiegelSvensson
 from .rates import (
